Add files via upload
Working test of heatmap image creation. Will generate an arbitrary-size image of a green-yellow-red image that scales diagonally from green to red from lower left to upper right corner of the image. Intended to be used for generating a background of whatever size is required then superimposing bubbles, labels, etc on top.